feat: update battle system

+++ Assets/Scripts/SOCharacterData.cs
@@ -14,11 +14,11 @@ public class SOCharacterData : ScriptableObject
     public CHARACTER_TYPE characterType;
     public string characterDescription;
     // public Sprite characterSprite;
-    public int health;
-    public int maxHealth;
+    public double health;
+    public double maxHealth;
 
     // For enemy only, 0 if is player
-    public int attackDamage; 
+    public double attackDamage; 
 
     public GameObject characterPrefab;
